% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_date.R
\name{add_date}
\alias{add_date}
\title{Add date/time column to data.frame}
\usage{
add_date(df)
}
\arguments{
\item{df}{Data frame containing columns `Year`, `Month`, `Day` and - optionally - `Hour`, `Minute` and/or `Second`}
}
\value{
data.frame consisting of the df input and a new column `Date`
}
\description{
Takes the `Year`, `Month`, `Day` and, if available, `Hour`, `Minute` and `Second` columns of a data.frame and uses them to produce a `Date` column that uses R's standard data/time format.
}
\details{
Converts YEARMODA to R date
}
\examples{


add_date(KA_weather)
add_date(Winters_hours_gaps)

 
}
\author{
Eike Luedeling
}
\keyword{utility}
