% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_phenologyFit.R
\name{c.bootstrap_phenologyFit}
\alias{c.bootstrap_phenologyFit}
\title{Concatenate bootstrap_phenologyfit objects}
\usage{
\method{c}{bootstrap_phenologyFit}(...)
}
\arguments{
\item{...}{Zero or multiple objects of type `bootstrap_phenologyfit`.}
}
\value{
An object of class `bootstrap_phenologyFit`, the concatenation of the
list of input object.
}
\description{
Concatenate bootstrap_phenologyfit objects
}
