% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyFitter.R
\name{chifull}
\alias{chifull}
\title{chifull}
\usage{
chifull(par, modelfn, bloomJDays, SeasonList, na_penalty = 365, ...)
}
\arguments{
\item{par}{numeric. vector of fit parameters}

\item{modelfn}{function. model function}

\item{bloomJDays}{numeric. vector of bloom hours! per year}

\item{SeasonList}{list. list of index vectors per year.}

\item{na_penalty}{numeric. penalty value for the residual if the model returns `NA`.}

\item{...}{further parameters to pass on to `modelfn`.}
}
\description{
RSS to minimise by `phenologyFitter`
}
\details{
function to compute the RSS
}
