% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnifiedModel.R
\name{ChuineCF}
\alias{ChuineCF}
\title{ChuineCF}
\usage{
ChuineCF(x, a, b, c)
}
\arguments{
\item{x}{temperature}

\item{a}{numeric. paramter}

\item{b}{numeric. paramter}

\item{c}{numeric. paramter}
}
\value{
Returns a numeric vector.
}
\description{
chilling and forcing response function for
the unified model by Chuine
}
\references{
Isabelle Chuine, A Unified Model for Budburst of Trees, J. theor. Biol. (2000) 207
}
