% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnifiedModel.R
\name{ChuineFstar}
\alias{ChuineFstar}
\title{ChuineFstar}
\usage{
ChuineFstar(Ctot, w, k)
}
\arguments{
\item{Ctot}{numeric. total state of chilling}

\item{w}{numeric > 0. parameter.}

\item{k}{numeric < 0. parameter.}
}
\value{
Returns a numeric vector.
}
\description{
Critical forcing value
}
\references{
Isabelle Chuine, A Unified Model for Budburst of Trees, J. theor. Biol. (2000) 207
}
