% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Date2YEARMODA.R
\name{Date2YEARMODA}
\alias{Date2YEARMODA}
\title{Date to YEARMODA conversion}
\usage{
Date2YEARMODA(Date, hours = FALSE)
}
\arguments{
\item{Date}{Date in R date format}

\item{hours}{boolean variable indicating whether YEARMODAHO should be
calculated (YEARMODA + hours)}
}
\value{
YEARMODA object (e.g. 20111224 for 24th December 2011)
}
\description{
Converts R dates to YEARMODA format
}
\details{
Converts R date to YEARMODA
}
\examples{


Date2YEARMODA(YEARMODA2Date(20001205))
Date2YEARMODA(YEARMODA2Date(19901003))

 
}
\author{
Eike Luedeling
}
\keyword{utility}
