% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSeason.R
\name{genSeason}
\alias{genSeason}
\title{Generate Seasons}
\usage{
genSeason(temps, mrange = c(8, 6), years)
}
\arguments{
\item{temps}{list. generated by chillR}

\item{mrange}{numeric. vector with two entries for the range of months (start month and end month)}

\item{years}{numeric. vector of years to be considered (with each entry specifying the year, in which the season **ends**)}
}
\description{
Identify the hours, days or months in a (monthly, daily or hourly) temperature dataset that belong to a particular season.
Seasons are defined according to the `mrange` argument, which specifies the start and end month of
the season. The `years` argument specifies the year, in which the dormancy season of interest ends.
}
