% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSeason.R
\name{genSeasonList}
\alias{genSeasonList}
\title{genSeasonList}
\usage{
genSeasonList(temps, mrange = c(8, 6), years)
}
\arguments{
\item{temps}{data.frame. Must have columns `Temp` containing the temperatures, `JDay`
the JDays, `Month` the months and `Year` the years. This
kind of data frame is for instance generated by \link{stack_hourly_temps}, but can also
be generated by hand or using a different routine.}

\item{mrange}{numeric. vector of length two for the range of months the season
should span. E.g. `mrange=c(8,6)` would span a season from August to next June.
There must not be any overlap in months, i.e. mrange[1] must be larger mrange[2].}

\item{years}{numeric. vector of years to be considered}
}
\value{
Returns a list of data frames. Each element of the list corresponds to one season. The
`data.frame` for each year has named columns `Temp`, `JDay` and `Year`.
}
\description{
Generates a list with data.frame elements for each season.
}
