\name{leap_year}

\alias{leap_year}

\title{Leap year finder}

\description{This function determines whether a given year is a leap year}

\usage{leap_year(x)}

\arguments{
  \item{x}{integer value, representing year number}
}

\details{Takes a year number as input, and returns TRUE if this is a leap year, and FALSE if not}

\value{ boolean variable (TRUE or FALSE)}

\references{https://en.wikipedia.org/wiki/Leap_year}

\author{Eike Luedeling, but based on pseudocode from Wikipedia}

\examples{

leap_year(2015)  
leap_year(2016)

 }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ utility }

