% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_temperature_scenarios.R
\name{load_ClimateWizard_scenarios}
\alias{load_ClimateWizard_scenarios}
\title{Load climate wizard scenarios}
\usage{
load_ClimateWizard_scenarios(path, prefix)
}
\arguments{
\item{path}{character string indicating the file path where the files are to be written.}

\item{prefix}{character string specifying the prefix for all files.}
}
\value{
a list of temperature scenarios.
}
\description{
This is a slightly modified version of the load_temperature_scenarios function that can
load climate scenarios downloaded with the getClimateWizardData and saved with the
save_temperature_scenarios function. This separate function is necessary, because the
climate scenarios are expressed as lists, with one element being a data.frame.
}
\examples{

temps<-list(Element1=data.frame(a=1,b=2),Element2=data.frame(a=c(2,3),b=c(8,4)))
# save_temperature_scenarios(temps,path=getwd(),prefix="temperatures")
# temps_reloaded<-load_temperature_scenarios(path=getwd(),prefix="temperatures")

 
}
\author{
Eike Luedeling
}
\keyword{utility}
