% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_JDay.R
\name{make_JDay}
\alias{make_JDay}
\title{Make Julian Day in dataframe}
\usage{
make_JDay(dateframe)
}
\arguments{
\item{dateframe}{data.frame, which should contain date information specified
as columns "Day", "Month" and "Year"}
}
\value{
Returns the same data.frame, but with column "JDay" added. This then
contains the Julian Dates.
}
\description{
This function produced Julian Dates (days of the year) from columns "Day",
"Month" and "Year" in a dataframe.
}
\examples{


dates<-data.frame(Year=c(1977,1980,2004,2011,2016),Month=c(11,8,3,12,8),Day=c(1,21,2,24,2)) 
make_JDay(dates)


}
\references{
The chillR package:
}
\author{
Eike Luedeling
}
\keyword{utilities}
