% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyFitter.R
\name{phenologyFit}
\alias{phenologyFit}
\title{phenologyFit}
\usage{
phenologyFit()
}
\value{
an empty object of class `phenologyFit`. It contains the named elements `model_fit` with 
the returned object from GenSA, `par` the best fit parameters, `pbloomJDays` the
predicted bloom JDays and the inputs
`par.guess`, `modelfn`, `bloomJDays`, and `SeasonList`. They are
all set to `NULL` by this function.
}
\description{
Constructor for class `phenologyFit`
}
