% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_phenologyFit.R
\name{plot.bootstrap_phenologyFit}
\alias{plot.bootstrap_phenologyFit}
\title{plot bootstrap_phenologyFit}
\usage{
\method{plot}{bootstrap_phenologyFit}(
  x,
  ylim = c(0.9 * min(c(x$object$bloomJDays, x$object$pbloomJDays)), 1.1 *
    max(c(x$object$bloomJDays, x$object$pbloomJDays))),
  ...
)
}
\arguments{
\item{x}{object of class `bootstrap_phenologyFit` to plot.}

\item{ylim}{numeric vector of length 2 with the limit for the y-axis}

\item{...}{additional graphical parameters to pass on.}
}
\value{
No return value.
}
\description{
Generic function to plot a `bootstrap_phenologyFit` object
}
