% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyFitter.R
\name{plot.phenologyFit}
\alias{plot.phenologyFit}
\title{plot phenologyFit}
\usage{
\method{plot}{phenologyFit}(
  x,
  ylim = c(0.9 * min(c(x$bloomJDays, x$pbloomJDays)), 1.1 * max(c(x$bloomJDays,
    x$pbloomJDays))),
  ...
)
}
\arguments{
\item{x}{object of class `phenologyFit` to plot.}

\item{ylim}{numeric vector of length 2 with the limit for the y-axis}

\item{...}{additional graphical parameters to pass on.}
}
\value{
No return value.
}
\description{
Generic function to plot a `phenologyFit` object
}
