% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getClimateWizard_scenarios.R
\name{plot_climateWizard_scenarios}
\alias{plot_climateWizard_scenarios}
\title{Plot mutltiple ClimateWizard scenarios obtained with getClimateWizard_scenarios}
\usage{
plot_climateWizard_scenarios(
  getscenarios_element,
  low_filter = -1000,
  high_filter = 1000,
  color = "cadetblue"
)
}
\arguments{
\item{getscenarios_element}{outputs from the getClimateWizard_scenarios function}

\item{low_filter}{numeric value specifying the lowest plausible value for the
variable of interest. This is sometimes necessary to exclude erroneous values
in the ClimateWizard database.}

\item{high_filter}{numeric value specifying the highest plausible value for the
variable of interest. This is sometimes necessary to exclude erroneous values
in the ClimateWizard database.}

\item{color}{color to be used for the plots.}
}
\value{
returns nothing, but a plot is produced as a side effect.
}
\description{
This function plots multiple scenarios obtained with the getClimateWizard_scenarios
function.
}
\examples{

#example is #d out, because of runtime issues.
#getC<-getClimateWizard_scenarios(coordinates=c(longitude=6.99,latitude=50.62),
#                                scenarios=c("rcp45","rcp45","rcp85","rcp85"),
#                                start_years=c(2035,2070,2035,2070),
#                                end_years=c(2065,2100,2065,2100),
#                                metric=c("monthly_tmean"),
#                                GCMs=c("all"))
#plot_climateWizard_scenarios(getC,low_filter=-6,high_filter=6,color="red")

}
\references{
Girvetz E, Ramirez-Villegas J, Navarro C, Rodriguez C, Tarapues J, undated.
ClimateWizard REST API for querying climate change data.
https://github.com/CIAT-DAPA/climate_wizard_api
}
\author{
Eike Luedeling
}
\keyword{utility}
