% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_phenologyFit.R
\name{predict.bootstrap_phenologyFit}
\alias{predict.bootstrap_phenologyFit}
\title{predict bootstrap_phenologyFit}
\usage{
\method{predict}{bootstrap_phenologyFit}(object, SeasonList, ...)
}
\arguments{
\item{object}{object of class `phenologyFit` to predict.}

\item{SeasonList}{List with data frames per season, see
\link{phenologyFit} for more details.}

\item{...}{additional parameters, ignored here}
}
\value{
A data.frame with one column `pbloomJDays` and a second one `Err`.
}
\description{
Generic function to predict a `bootstrap_phenologyFit` object.
}
