% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyFitter.R
\name{predict.phenologyFit}
\alias{predict.phenologyFit}
\title{predict phenologyFit}
\usage{
\method{predict}{phenologyFit}(object, SeasonList, ...)
}
\arguments{
\item{object}{object of class `phenologyFit` to predict.}

\item{SeasonList}{List with data frames per season, see
\link{phenologyFit} for more details.}

\item{...}{additional parameters, ignored here}
}
\value{
A numeric vector is returned with a predicted bloom day per
Season in `SeasonList`. If `SeasonList` is missing, the
original `SeasonList` is used for prediction.
}
\description{
Generic function to predict a `phenologyFit` object.
}
