% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyFitter.R
\name{print.phenologyFit}
\alias{print.phenologyFit}
\title{print phenologyFit}
\usage{
\method{print}{phenologyFit}(x, ...)
}
\arguments{
\item{x}{class phenologyFit. object to print}

\item{...}{additional parameters, ignored here}
}
\value{
No return value.
}
\description{
print phenologyFit
}
