% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnifiedModel.R
\name{StepChill_Wrapper}
\alias{StepChill_Wrapper}
\title{StepChill_Wrapper}
\usage{
StepChill_Wrapper(x, par)
}
\arguments{
\item{x}{data.frame with at least columns `Temp` and `JDay`}

\item{par}{numeric vector of length 7 with the parameters of the
StepChill model: 1. Tc, 2. bf, 3. cf, 4. Cstar and 5. Fstar.}
}
\value{
A single numeric value with the JDay prediction for the
temperaturs in `x$Temp` and the model parameters
in `par`.
}
\description{
Same as UniChill_Wrapper, but with a step function for chilling
}
\references{
Isabelle Chuine, A Unified Model for Budburst of Trees, J. theor. Biol. (2000) 207

Asse et al., Process-based models outcompete correlative models in projecting spring
phenology of trees in a future warmer climate,Agricultural and Forest Meteorology (2020) 107913
}
\author{
Carsten Urbach <urbach@hiskp.uni-bonn.de>
}
