% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_phenologyFit.R
\name{summary.bootstrap_phenologyFit}
\alias{summary.bootstrap_phenologyFit}
\title{summary.bootstrap_phenologyFit}
\usage{
\method{summary}{bootstrap_phenologyFit}(object, ...)
}
\arguments{
\item{object}{class `bootstrap_phenologyFit` to summarise}

\item{...}{generic parameters, ignored here}
}
\value{
No return value.
}
\description{
Summarise a `bootstrap_phenologyFit` object
}
