% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyFitter.R
\name{summary.phenologyFit}
\alias{summary.phenologyFit}
\title{summary phenologyFit}
\usage{
\method{summary}{phenologyFit}(object, ...)
}
\arguments{
\item{object}{class phenologyFit. object to summarise}

\item{...}{additional parameters, ignored here}
}
\value{
No return value.
}
\description{
summary phenologyFit
}
