% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnifiedModel.R
\name{UniChill_Wrapper}
\alias{UniChill_Wrapper}
\title{UniChill_Wrapper}
\usage{
UniChill_Wrapper(x, par)
}
\arguments{
\item{x}{data.frame with at least columns `Temp` and `JDay`}

\item{par}{numeric vector of length 7 with the parameters of the
UniChill model: 1. ac, 2. bc, 3. cc, 4. bf, 5. cf, 6. Cstar and 7. Fstar.}
}
\value{
A single numeric value with the JDay prediction for the
temperaturs in `x$Temp` and the model parameters
in `par`.
}
\description{
UniChill_Wrapper
}
\references{
Isabelle Chuine, A Unified Model for Budburst of Trees, J. theor. Biol. (2000) 207
}
\author{
Carsten Urbach <urbach@hiskp.uni-bonn.de>
}
