% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnifiedModel.R
\name{UnifiedModel_Wrapper}
\alias{UnifiedModel_Wrapper}
\title{UnifiedModel_Wrapper}
\usage{
UnifiedModel_Wrapper(x, par)
}
\arguments{
\item{x}{data.frame with at least columns `Temp` and `JDay`}

\item{par}{numeric vector of length 9 with the parameters of the
unified model: 1. ac, 2. bc, 3. cc, 4. bf, 5. cf, 6. w, 7. k,
8. Cstar and 9. tc.}
}
\value{
A single numeric value with the JDay prediction for the
temperaturs in `x$Temp` and the Unified Model parameters
in `par`.
}
\description{
UnifiedModel_Wrapper
}
\references{
Isabelle Chuine, A Unified Model for Budburst of Trees, J. theor. Biol. (2000) 207
}
\author{
Carsten Urbach <urbach@hiskp.uni-bonn.de>
}
