% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnifiedModel.R
\name{UniForce_Wrapper}
\alias{UniForce_Wrapper}
\title{UniForce_Wrapper}
\usage{
UniForce_Wrapper(x, par)
}
\arguments{
\item{x}{data.frame with at least columns `Temp` and `JDay`}

\item{par}{numeric vector of length 4 with the parameters of the
UniForce model: 1. bf, 2. cf, 3. Fstar, 4. t1.}
}
\value{
A single numeric value with the JDay prediction for the
temperaturs in `x$Temp` and the Unified Model parameters
in `par`.
}
\description{
UniForce_Wrapper
}
\references{
Isabelle Chuine, A Unified Model for Budburst of Trees, J. theor. Biol. (2000) 207
}
\author{
Carsten Urbach <urbach@hiskp.uni-bonn.de>
}
