\name{YEARMODA2Date}

\alias{YEARMODA2Date}

\title{YEARMODA to Date conversion}

\description{Converts dates in YEARMODA format to R date format}

\usage{YEARMODA2Date(YEARMODA)}

\arguments{
  \item{YEARMODA}{Date in YEARMODA format (e.g. 20160206 for 6th Feb 2016)}
}

\details{Converts YEARMODA to R date}

\value{ Date object}

\author{Eike Luedeling}

\examples{

YEARMODA2Date(20001205)
YEARMODA2Date(19901003)

 }
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ utility }

