% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chillR-package.R
\docType{data}
\name{KA_bloom}
\alias{KA_bloom}
\title{Cherry bloom data for Klein-Altendorf, Germany}
\format{
A data frame with the following 2 variables.  \describe{
\item{Year}{a numeric vector, indicating the observation year}
\item{pheno}{ a vector that, when coerced by as.numeric, contains
bloom data in Julian dates (day of the year)} }
}
\source{
data were collected by Achim Kunz and Michael Blanke, University of
Bonn
}
\description{
Bloom data of sweet cherry var. 'Schneiders spaete Knorpelkirsche' recorded
at Klein-Altendorf, Germany, the experimental station of the University of
Bonn
}
\examples{

data(KA_bloom)

}
\references{
Luedeling E, Kunz A and Blanke M, 2013. Identification of
chilling and heat requirements of cherry trees - a statistical approach.
International Journal of Biometeorology 57,679-689.
}
\keyword{datasets}
