% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chillR-package.R
\docType{data}
\name{KA_weather}
\alias{KA_weather}
\title{Weather data for Klein-Altendorf, Germany}
\format{
A data frame with observations on the following 5 variables.
\describe{ \item{Year}{a numeric vector - the observation year}
\item{Month}{a numeric vector - the observation month}
\item{Day}{a numeric vector - the observation day}
\item{Tmax}{a numeric vector - daily maximum temperature}
\item{Tmin}{a numeric vector - daily minimum temperature} }
}
\source{
data were collected by Achim Kunz and Michael Blanke, University of
Bonn
}
\description{
Daily temperature data from Klein-Altendorf, Germany, for use in combination
with the example phenology dataset KA_bloom.
}
\examples{

data(KA_weather)

}
\references{
Luedeling E, Kunz A and Blanke M, 2013. Identification of
chilling and heat requirements of cherry trees - a statistical approach.
International Journal of Biometeorology 57,679-689.
}
\keyword{datasets}
