% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leap_year.R
\name{leap_year}
\alias{leap_year}
\title{Leap year finder}
\usage{
leap_year(x)
}
\arguments{
\item{x}{integer value, representing year number}
}
\value{
boolean variable (TRUE or FALSE)
}
\description{
This function determines whether a given year is a leap year
}
\details{
Takes a year number as input, and returns TRUE if this is a leap year, and
FALSE if not
}
\examples{


leap_year(2015)  
leap_year(2016)

 
}
\references{
https://en.wikipedia.org/wiki/Leap_year
}
\author{
Eike Luedeling, but based on pseudocode from Wikipedia
}
\keyword{utility}
