% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runn_mean.R
\name{runn_mean}
\alias{runn_mean}
\title{Running mean of a vector}
\usage{
runn_mean(vec, runn_mean)
}
\arguments{
\item{vec}{numeric vector}

\item{runn_mean}{number of vector elements to use for calculating the
running mean}
}
\value{
numeric vector containing the running mean
}
\description{
Function to calculate the running mean of a numeric vector
}
\examples{


plot(runn_mean(rnorm(1000),150))

}
\author{
Eike Luedeling
}
\keyword{~kwd1}
\keyword{~kwd2}

