% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YEARMODA2Date.R
\name{YEARMODA2Date}
\alias{YEARMODA2Date}
\title{YEARMODA to Date conversion}
\usage{
YEARMODA2Date(YEARMODA)
}
\arguments{
\item{YEARMODA}{Date in YEARMODA format (e.g. 20160206 for 6th Feb 2016)}
}
\value{
Date object
}
\description{
Converts dates in YEARMODA format to R date format
}
\details{
Converts YEARMODA to R date
}
\examples{


YEARMODA2Date(20001205)
YEARMODA2Date(19901003)

 
}
\author{
Eike Luedeling
}
\keyword{utility}
